Bootstrapping in a generic host
