+++ src/AspNetCoreIntegrations/AspNetCoreIntegrations.csproj
@@ -7,7 +7,8 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Bootstrapping\Bootstrapping.csproj" />
+      <ProjectReference Include="..\ServiceContainer.Drivers\ServiceContainer.Drivers.csproj" />
+      <ProjectReference Include="..\HostingIntegrations\HostingIntegrations.csproj" />
       <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
       <ProjectReference Include="..\Foundation\Foundation.csproj" />
     </ItemGroup>
