+++ src/Bootstrapping/Bootstrapping.csproj
@@ -3,22 +3,13 @@
     <Import Project="../MSBuild/Packable.props"/>
 
     <ItemGroup Label="Package references">
-        <PackageReference Include="Humanizer.Core" GeneratePathProperty="true" PrivateAssets="all" />
         <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
+        <PackageReference Include="Microsoft.Extensions.Hosting" />
     </ItemGroup>
 
     <ItemGroup Label="Project references">
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+        <ProjectReference Include="..\ServiceContainer.Drivers\ServiceContainer.Drivers.csproj" />
+        <ProjectReference Include="..\HostingIntegrations\HostingIntegrations.csproj" />
         <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
-        <ProjectReference Include="..\SourceGenerators.BootstrapDrivers\SourceGenerators.BootstrapDrivers.csproj"
-                          OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
-    </ItemGroup>
-
-    <ItemGroup Label="Transitive bootstrap driver generator">
-        <None Include="..\SourceGenerators.BootstrapDrivers\bin\$(Configuration)\netstandard2.0\IOKode.OpinionatedFramework.SourceGenerators.BootstrapDrivers.dll"
-              Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
-        <None Include="$(PKGHumanizer_Core)\lib\netstandard2.0\*.dll"
-              Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
     </ItemGroup>
 </Project>
