+++ src/Bootstrapping/OpinionatedFrameworkBootstrapping.cs
@@ -1,47 +1,19 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
+using IOKode.OpinionatedFramework.HostingIntegrations;
 using IOKode.OpinionatedFramework.ServiceContainer;
+using IOKode.OpinionatedFramework.ServiceContainer.Drivers;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 using System.Threading;
 using System.Threading.Tasks;
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
-using IOKode.OpinionatedFramework.ServiceLocation;
+using Microsoft.Extensions.Hosting;
 
 namespace IOKode.OpinionatedFramework.Bootstrapping;
 
 /// <summary>
-/// Selects configured drivers, validates them as a group, registers their services, and initializes lifecycle tasks.
+/// Initializes the framework container or creates and starts a framework host.
 /// </summary>
 public static class OpinionatedFrameworkBootstrapping
 {
-    /// <summary>Selects, validates, and registers drivers using the <c>OpinionatedFramework</c> configuration section.</summary>
-    /// <param name="configuration">The host configuration used to select and configure drivers.</param>
-    /// <exception cref="ArgumentNullException"><paramref name="configuration"/> is <see langword="null"/>.</exception>
-    /// <exception cref="BootstrapConfigurationException">Driver selection or validation fails.</exception>
-    public static void RegisterDrivers(IConfiguration configuration)
-    {
-        ArgumentNullException.ThrowIfNull(configuration);
-
-        var frameworkConfiguration = configuration.GetSection("OpinionatedFramework");
-        var sharedState = new Dictionary<string, object>(StringComparer.Ordinal);
-        var selectedDrivers = SelectDrivers(configuration, frameworkConfiguration, sharedState);
-
-        var validationErrors = selectedDrivers
-            .SelectMany(selectedDriver => selectedDriver.Descriptor.Validate(selectedDriver.Context).Errors)
-            .ToArray();
-        if (validationErrors.Length > 0)
-        {
-            throw new BootstrapConfigurationException(validationErrors);
-        }
-
-        foreach (var selectedDriver in selectedDrivers)
-        {
-            selectedDriver.Descriptor.Register(selectedDriver.Context);
-        }
-    }
-
     /// <summary>
     /// Initialize the OpinionatedFramework container.
     /// </summary>
@@ -50,140 +22,32 @@ public static class OpinionatedFrameworkBootstrapping
         Container.Initialize();
     }
 
-    /// <summary>Executes lifecycle startup tasks after the container has been initialized.</summary>
-    /// <param name="cancellationToken">Cancels lifecycle startup.</param>
-    /// <returns>A handle that owns the initialized container.</returns>
-    public static async Task<IAsyncDisposable> StartLifecycleAsync(CancellationToken cancellationToken = default)
+    /// <summary>Creates and starts an <see cref="IHost"/> using the configured framework drivers.</summary>
+    /// <param name="configuration">The host configuration used to select and configure drivers.</param>
+    /// <param name="cancellationToken">Cancels host startup.</param>
+    /// <returns>A handle that owns the host and initialized container.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="configuration"/> is <see langword="null"/>.</exception>
+    /// <exception cref="BootstrapConfigurationException">Driver selection or validation fails.</exception>
+    public static async Task<HostHandle> StartAsync(IConfiguration configuration,
+        CancellationToken cancellationToken = default)
     {
+        ArgumentNullException.ThrowIfNull(configuration);
+
+        var host = Host.CreateDefaultBuilder()
+            .ConfigureAppConfiguration(configurationBuilder => configurationBuilder.AddConfiguration(configuration))
+            .ConfigureServices((hostContext, _) => DriverRegistration.RegisterDrivers(hostContext.Configuration))
+            .UseServiceProviderFactory(new OpinionatedFrameworkServiceProviderFactory())
+            .Build();
+
         try
         {
-            var startupTasks = Locator.ServiceProvider!.GetServices<IStartupTask>();
-            foreach (var startupTask in startupTasks)
-            {
-                await startupTask.StartAsync(cancellationToken);
-            }
-
-            return new ContainerHandle();
+            await host.StartAsync(cancellationToken);
+            return new HostHandle(host);
         }
         catch
         {
-            await Container.Advanced.DisposeAsync();
+            host.Dispose();
             throw;
         }
     }
-    
-    /// <summary>Registers drivers, initializes the service container, and executes startup tasks.</summary>
-    /// <param name="configuration">The host configuration used to select and configure drivers.</param>
-    /// <param name="cancellationToken">Cancels framework startup tasks.</param>
-    /// <returns>A handle that owns the initialized container.</returns>
-    /// <exception cref="ArgumentNullException"><paramref name="configuration"/> is <see langword="null"/>.</exception>
-    /// <exception cref="BootstrapConfigurationException">Driver selection or validation fails.</exception>
-    public static async Task<IAsyncDisposable> StartAsync(IConfiguration configuration,
-        CancellationToken cancellationToken = default)
-    {
-        RegisterDrivers(configuration);
-        InitializeContainer();
-        return await StartLifecycleAsync(cancellationToken);
-    }
-
-    private static IReadOnlyCollection<SelectedDriver> SelectDrivers(IConfiguration configuration,
-        IConfigurationSection frameworkConfiguration, IDictionary<string, object> sharedState)
-    {
-        var selectedDrivers = new List<SelectedDriver>();
-        foreach (var contractDrivers in BootstrapDriverCatalog.RegisteredDrivers.GroupBy(driver => driver.ContractType))
-        {
-            var configurationKeys = contractDrivers
-                .Select(driver => driver.ConfigurationKey)
-                .Distinct(StringComparer.OrdinalIgnoreCase)
-                .ToArray();
-            var namedInstanceModes = contractDrivers
-                .Select(driver => driver.SupportsNamedInstances)
-                .Distinct()
-                .ToArray();
-
-            if (configurationKeys.Length != 1 || namedInstanceModes.Length != 1)
-            {
-                throw new BootstrapConfigurationException(
-                    $"Drivers for contract '{contractDrivers.Key.FullName}' must use the same configuration key and instance mode.");
-            }
-
-            var configurationSection = frameworkConfiguration.GetSection(configurationKeys[0]);
-            if (namedInstanceModes[0])
-            {
-                foreach (var instanceSection in configurationSection.GetChildren())
-                {
-                    selectedDrivers.Add(SelectConfiguredDriver(contractDrivers, configuration, frameworkConfiguration,
-                        instanceSection, instanceSection.Key, sharedState));
-                }
-
-                continue;
-            }
-
-            var configuredDriverKey = configurationSection["Driver"];
-            if (string.IsNullOrWhiteSpace(configuredDriverKey))
-            {
-                var defaultDrivers = contractDrivers.Where(driver => driver.IsDefault).ToArray();
-                if (defaultDrivers.Length == 0)
-                {
-                    continue;
-                }
-
-                if (defaultDrivers.Length > 1)
-                {
-                    throw new BootstrapConfigurationException(
-                        $"Multiple default drivers are registered for contract '{contractDrivers.Key.FullName}'.");
-                }
-
-                selectedDrivers.Add(CreateSelectedDriver(defaultDrivers[0], configuration, frameworkConfiguration,
-                    configurationSection, null, sharedState));
-                continue;
-            }
-
-            selectedDrivers.Add(SelectConfiguredDriver(contractDrivers, configuration, frameworkConfiguration,
-                configurationSection, null, sharedState));
-        }
-
-        return selectedDrivers;
-    }
-
-    private static SelectedDriver SelectConfiguredDriver(IEnumerable<BootstrapDriverDescriptor> availableDrivers,
-        IConfiguration configuration, IConfigurationSection frameworkConfiguration,
-        IConfigurationSection driverConfiguration, string? instanceName, IDictionary<string, object> sharedState)
-    {
-        var configuredDriverKey = driverConfiguration["Driver"];
-        if (string.IsNullOrWhiteSpace(configuredDriverKey))
-        {
-            throw new BootstrapConfigurationException(
-                $"Configuration '{driverConfiguration.Path}:Driver' is required.");
-        }
-
-        var availableDriversArray = availableDrivers.ToArray();
-        var selectedDriver = availableDriversArray.SingleOrDefault(driver =>
-            string.Equals(driver.DriverKey, configuredDriverKey, StringComparison.OrdinalIgnoreCase));
-        if (selectedDriver is null)
-        {
-            var availableKeys = string.Join(", ", availableDriversArray.Select(driver => driver.DriverKey).Order());
-            throw new BootstrapConfigurationException(
-                $"Unknown driver '{configuredDriverKey}' for contract '{availableDriversArray[0].ContractType.FullName}'. " +
-                $"Available drivers: {availableKeys}. Configuration path: {driverConfiguration.Path}:Driver.");
-        }
-
-        return CreateSelectedDriver(selectedDriver, configuration, frameworkConfiguration, driverConfiguration,
-            instanceName, sharedState);
-    }
-
-    private static SelectedDriver CreateSelectedDriver(BootstrapDriverDescriptor descriptor,
-        IConfiguration configuration, IConfigurationSection frameworkConfiguration,
-        IConfigurationSection driverConfiguration, string? instanceName, IDictionary<string, object> sharedState)
-    {
-        return new SelectedDriver(descriptor, new BootstrapDriverContext(
-            Container.Services,
-            configuration,
-            frameworkConfiguration,
-            driverConfiguration,
-            instanceName,
-            sharedState));
-    }
-
-    private sealed record SelectedDriver(BootstrapDriverDescriptor Descriptor, BootstrapDriverContext Context);
 }
