+++ src/ContractImplementations.Aes256GcmModeEncrypter/BootstrapDriver.cs
@@ -1,6 +1,5 @@
 using System;
-using IOKode.OpinionatedFramework.Bootstrapping;
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
+using IOKode.OpinionatedFramework.Drivers.Abstractions;
 using IOKode.OpinionatedFramework.Security;
 using Microsoft.Extensions.DependencyInjection;
