+++ src/ContractImplementations.CommandExecutor/ContractImplementations.CommandExecutor.csproj
@@ -7,7 +7,7 @@
     </ItemGroup>
 
     <ItemGroup Label="Project references">
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+        <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj" />
         <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
         <ProjectReference Include="..\Foundation\Foundation.csproj"/>
         <ProjectReference Include="..\Internals\Internals.csproj"/>
