+++ src/ContractImplementations.Events/BootstrapDriver.cs
@@ -1,5 +1,4 @@
-using IOKode.OpinionatedFramework.Bootstrapping;
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
+using IOKode.OpinionatedFramework.Drivers.Abstractions;
 using IOKode.OpinionatedFramework.Events;
 
 [assembly: BootstrapDriver<IEventDispatcher,
