+++ src/ContractImplementations.GoogleCloudStorage/ContractImplementations.GoogleCloudStorage.csproj
@@ -3,7 +3,7 @@
     <Import Project="../MSBuild/Packable.props"/>
 
     <ItemGroup Label="Project references">
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj"/>
+        <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj"/>
         <ProjectReference Include="..\ContractImplementations.FileSystem\ContractImplementations.FileSystem.csproj"/>
         <ProjectReference Include="..\Foundation\Foundation.csproj"/>
     </ItemGroup>
