+++ src/ContractImplementations.Hangfire/ContractImplementations.Hangfire.csproj
@@ -3,13 +3,14 @@
     <Import Project="../MSBuild/Packable.props"/>
 
     <ItemGroup>
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+        <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj" />
         <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
         <ProjectReference Include="..\Foundation\Foundation.csproj" />
     </ItemGroup>
 
     <ItemGroup>
         <PackageReference Include="Hangfire.Core" />
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
     </ItemGroup>
 
 </Project>
