+++ src/ContractImplementations.Hangfire/Jobs/BootstrapDrivers.cs
@@ -1,5 +1,4 @@
-using IOKode.OpinionatedFramework.Bootstrapping;
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
+using IOKode.OpinionatedFramework.Drivers.Abstractions;
 using IOKode.OpinionatedFramework.Jobs;
 
 [assembly: BootstrapDriver<IJobEnqueuer,
