+++ src/ContractImplementations.LocalFileSystem/BootstrapDriver.cs
@@ -1,6 +1,5 @@
-using IOKode.OpinionatedFramework.Bootstrapping;
 using System.Collections.Generic;
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
+using IOKode.OpinionatedFramework.Drivers.Abstractions;
 using IOKode.OpinionatedFramework.FileSystem;
 using Microsoft.Extensions.DependencyInjection;
