+++ src/ContractImplementations.MicrosoftConfiguration/ContractImplementations.MicrosoftConfiguration.csproj
@@ -3,7 +3,7 @@
     <Import Project="../MSBuild/Packable.props"/>
 
     <ItemGroup Label="Project dependencies">
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+        <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj" />
         <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
         <ProjectReference Include="..\Foundation\Foundation.csproj"/>
     </ItemGroup>
