+++ src/ContractImplementations.MicrosoftLogging/BootstrapDriver.cs
@@ -1,4 +1,4 @@
-using IOKode.OpinionatedFramework.Bootstrapping.Abstractions;
+using IOKode.OpinionatedFramework.Drivers.Abstractions;
 using IOKode.OpinionatedFramework.Logging;
 using Microsoft.Extensions.Logging;
