+++ src/ContractImplementations.MicrosoftLogging/ContractImplementations.MicrosoftLogging.csproj
@@ -9,7 +9,7 @@
     </ItemGroup>
 
     <ItemGroup Label="Project dependencies">
-        <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+        <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj" />
         <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
         <ProjectReference Include="..\Foundation\Foundation.csproj"/>
     </ItemGroup>
