+++ src/ContractImplementations.TaskRunJobs/ContractImplementations.TaskRunJobs.csproj
@@ -3,8 +3,12 @@
     <Import Project="../MSBuild/Packable.props"/>
 
     <ItemGroup Label="Project references">
-      <ProjectReference Include="..\Bootstrapping.Abstractions\Bootstrapping.Abstractions.csproj" />
+      <ProjectReference Include="..\Drivers.Abstractions\Drivers.Abstractions.csproj" />
       <ProjectReference Include="..\ServiceContainer\ServiceContainer.csproj" />
       <ProjectReference Include="..\Foundation\Foundation.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
     </ItemGroup>
 </Project>
